requirements.txt
argon2-cffi==20.1.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.42.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-widget-tweaks==1.4.8
idna==2.10
jsonfield==3.1.0
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.10
whitenoise==5.2.0
djangorestframework~=3.11.0